refactor: replace `useTag*` hooks with plain Qraft hook calls
