+++ packages/sdk-react/src/components/payables/PayableDetails/PayableDetailsForm/usePayableDetailsForm.tsx
@@ -1,7 +1,7 @@
+import { useAPI } from '@/api/client';
 import { useCounterpartList } from '@/core/queries/useCounterpart';
 import { useCounterpartsAddresses } from '@/core/queries/useCounterpartsAddresses';
 import { useCounterpartsBankAccountsList } from '@/core/queries/useCouterpartsBankAccounts';
-import { useTagListQuery } from '@/core/queries/useTag';
 
 export type UsePayableDetailsFormProps = {
   currentCounterpartId: string;
@@ -10,7 +10,9 @@ export type UsePayableDetailsFormProps = {
 export function usePayableDetailsForm({
   currentCounterpartId,
 }: UsePayableDetailsFormProps) {
-  const tagQuery = useTagListQuery();
+  const { api } = useAPI();
+  const tagQuery = api.tags.getTags.useQuery({});
+
   const counterpartQuery = useCounterpartList();
   const counterpartAddressQuery =
     useCounterpartsAddresses(currentCounterpartId);
